% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{Create_SCST_Object}
\alias{Create_SCST_Object}
\title{Create SCST Object}
\usage{
Create_SCST_Object(
  reference,
  methods,
  group_label = NULL,
  batch_label = NULL,
  dilution_factor = NULL,
  volume = NULL,
  spatial_x = NULL,
  spatial_y = NULL,
  seed = random_seed()
)
}
\arguments{
\item{reference}{The reference gene expression matrix used for estimating parameters}

\item{methods}{The simulation methods used for estimating parameters and data simulation. Users can specify more than 1 methods to use and the procedure will check the prior information which are indispensable for each method.}

\item{group_label}{The group labels of cells/spots/bins}

\item{batch_label}{The batch labels of cells/spots/bins}

\item{dilution_factor}{If ERCC spike-in RNAs were added into the single-cell solution, the dilution factor can be used for estimating parameters from real using certain methods (such as SPARSim)}

\item{volume}{The volume of ERCC spike-in mix added into the single-cell solution}

\item{spatial_x}{The coordinates of the spots/bins on the x-axis for the spatial transcriptome data}

\item{spatial_y}{The coordinates of the spots/bins on the y-axis for the spatial transcriptome data}

\item{seed}{The random seed used for reproducibility}
}
\description{
Create SCST Object
}
