% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropsim.R
\name{dropsim_estimation}
\alias{dropsim_estimation}
\title{Estimation Parameters Using dropsim}
\usage{
dropsim_estimation(SCST_Object, verbose = FALSE)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}
}
\description{
Estimation Parameters Using dropsim
}
\references{
Github URL: <https://github.com/marchinilab/dropsim>
}
