% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntoy.R
\name{dyntoy_simulation}
\alias{dyntoy_simulation}
\title{Simulate ScRNA-seq Data Using dyntoy}
\usage{
dyntoy_simulation(
  estimated_result,
  cell_num = NULL,
  gene_num = NULL,
  de_prop = 0.2,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [SCST::dyntoy_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in dyntoy, see [dyntoy::generate_dataset()]}
}
\description{
Simulate ScRNA-seq Data Using dyntoy
}
