% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_parameters.R
\name{Estimate_parameters}
\alias{Estimate_parameters}
\title{Estimate Parameters From Reference Data}
\usage{
Estimate_parameters(SCST_Object, verbose = TRUE, ...)
}
\arguments{
\item{SCST_Object}{A SCST object for downstream procedure, which is generated by [SCST::Create_SCST_Object]}

\item{verbose}{Whether to return messages or not}

\item{...}{Other customed parameters represented in the function of methods}
}
\description{
Estimate Parameters From Reference Data
}
