% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_assosication_methods.R
\name{EvalAssociationMethods}
\alias{EvalAssociationMethods}
\title{Evaluation of Clustering Methods Based on Simulated Data
Evaluate Cell Association Measurements Using Simulated Data}
\usage{
EvalAssociationMethods(SimulationResult, n_cores = 1, verbose = TRUE)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated data.}

\item{n_cores}{The number of computational cores used in the clustering methods. Default is 1.}

\item{verbose}{Whether to return messages or not. Default is TRUE.}
}
\description{
Evaluation of Clustering Methods Based on Simulated Data
Evaluate Cell Association Measurements Using Simulated Data
}
\references{
Skinnider M A, Squair J W, Foster L J. Evaluating measures of association for single-cell transcriptomics[J]. Nature methods, 2019, 16(5): 381-386.
}
