% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_batch_removal_methods.R
\name{EvalBatchRemovalMethods}
\alias{EvalBatchRemovalMethods}
\title{Evaluation of Batch-Removal Methods Based on Simulated Data}
\usage{
EvalBatchRemovalMethods(
  SimulationResult,
  conda_env,
  min_cells = 0,
  min_genes = 0,
  min_total_counts = 0,
  min_counts_per_gene = 1,
  nFeatures = 2000,
  PCs = 20,
  nlayers = 2,
  nlatent = 20,
  genelikelihood = "nb",
  batch_size = 64,
  epochs = NULL,
  k_NNs = 20,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated data.}

\item{conda_env}{The path of conda environment. Additionally, the packages of scvi/scanpy/scanorama/scgen are strongly recommended to be installed by users.}

\item{min_cells}{Include genes detected in at least min_cells of cells. The genes detected below the number of min_cells are removed. Default no genes are removed.}

\item{min_genes}{Include cells where at least min_genes of genes are detected. The cells contain below the number of min_genes are removed. Default no genes are removed.}

\item{min_total_counts}{Include cells whose total counts are higher than min_total_counts.}

\item{min_counts_per_gene}{Include genes where more than min_counts_per_gene of counts are detected.}

\item{nFeatures}{The number of highly variable genes. Default is 2000.}

\item{PCs}{The number of principal components used. Default is 20 and the value higher than 10 is recommended.}

\item{nlayers}{Number of hidden layers used for encoder and decoder NNs. Default is 2.}

\item{nlatent}{The dimensionality of the latent space. Default is 20.}

\item{genelikelihood}{The distribution model for the data, which is one of nb (Negative binomial distribution), zinb (Zero-inflated negative binomial distribution) and poisson (Poisson distribution). Default is nb.}

\item{batch_size}{Number of samples per training batch. Default is 64.}

\item{epochs}{Number of training epochs. Default is NULL and it is determined automatically.}

\item{k_NNs}{An integer scalar specifying the number of nearest neighbors to consider when identifying MNNs.}

\item{seed}{The random seed used for reproducibility.}

\item{verbose}{Whether to return messages or not. Default is TRUE.}
}
\description{
Evaluation of Batch-Removal Methods Based on Simulated Data
}
