% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_clustering_methods.R
\name{EvalClusteringMethods}
\alias{EvalClusteringMethods}
\title{Evaluation of Clustering Methods Based on Simulated Data}
\usage{
EvalClusteringMethods(
  SimulationResult,
  min_cells = 0,
  min_genes = 0,
  min_total_counts = 0,
  min_counts_per_gene = 1,
  PCs = 20,
  n_cores = 1,
  learning_rate = 0.001,
  batch_size = 64,
  epochs = 200,
  tsne_dims = 3,
  perplexity = 30,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated data.}

\item{min_cells}{Include genes detected in at least min_cells of cells. The genes detected below the number of min_cells are removed. Default no genes are removed.}

\item{min_genes}{Include cells where at least min_genes of genes are detected. The cells contain below the number of min_genes are removed. Default no genes are removed.}

\item{min_total_counts}{Include cells whose total counts are higher than min_total_counts.}

\item{min_counts_per_gene}{Include genes where more than min_counts_per_gene of counts are detected.}

\item{PCs}{The used number of principal components used. Default is 20.}

\item{n_cores}{The number of computational cores used in the clustering methods. Default is 1.}

\item{learning_rate}{Learning rate for training the artificial neural network. Default is 0.001.}

\item{batch_size}{Number of samples per training batch. Default is 64.}

\item{epochs}{Number of training epochs. Default is 200.}

\item{tsne_dims}{Output dimensionality of TSNE. Default is 3.}

\item{perplexity}{Perplexity parameter used for TSNE. Default is 30.}

\item{seed}{The random seed used for reproducibility.}

\item{verbose}{Whether to return messages or not. Default is TRUE.}
}
\description{
Evaluation of Clustering Methods Based on Simulated Data
}
