% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_DEGs_detection_methods.R
\name{EvalDEAMethods}
\alias{EvalDEAMethods}
\title{Evaluate Differential Expression Analysis Methods Using Simulated Data}
\usage{
EvalDEAMethods(SimulationResult, verbose = TRUE)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated results.}

\item{verbose}{Whether the process massages are returned.}
}
\description{
Evaluate Differential Expression Analysis Methods Using Simulated Data
}
\references{
Soneson C, Robinson M D. Bias, robustness and scalability in single-cell differential expression analysis. Nature methods, 2018, 15(4): 255-261. <https://doi.org/10.1038/nmeth.4612>
}
