% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_domain_detection_methods.R
\name{EvalDomainDetectionMethods}
\alias{EvalDomainDetectionMethods}
\title{Evaluation of Domain Detection Methods Based on Simulated Spatial Data}
\usage{
EvalDomainDetectionMethods(
  SimulationResult,
  conda_env,
  min_cells = 0,
  min_genes = 0,
  min_total_counts = 0,
  min_counts_per_gene = 1,
  nFeatures = 2000,
  PCs = 20,
  learning_rate = 0.001,
  epochs = 1000,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated spatial data.}

\item{conda_env}{The path of conda environment. Additionally, the packages or the environment (Python packages: DeepST, STAGATE, GraphST and SCGDL) should be installed by users previously. The detail configuration in conda is demonstrated below.}

\item{min_cells}{Include genes detected in at least min_cells of cells. The genes detected below the number of min_cells are removed. Default no genes are removed.}

\item{min_genes}{Include cells where at least min_genes of genes are detected. The cells contain below the number of min_genes are removed. Default no genes are removed.}

\item{min_total_counts}{Include cells whose total counts are higher than min_total_counts.}

\item{min_counts_per_gene}{Include genes where more than min_counts_per_gene of counts are detected.}

\item{nFeatures}{The number of highly variable genes. Default is 2000.}

\item{PCs}{The number of principal components used. Default is 20.}

\item{learning_rate}{Learning rate for training the artificial neural network. Default is 0.001.}

\item{epochs}{Number of training epochs. Default is 1000.}

\item{seed}{The random seed used for reproducibility.}

\item{verbose}{Whether to return messages or not. Default is TRUE.}
}
\description{
Evaluation of Domain Detection Methods Based on Simulated Spatial Data
}
