% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval_clustering_methods.R
\name{EvaluateAppScenario}
\alias{EvaluateAppScenario}
\title{Evaluate Clustering Methods Using Simulated Data}
\usage{
EvaluateAppScenario(
  SimulationResult,
  eval_clustering = FALSE,
  eval_cell_association = FALSE,
  eval_annotation = FALSE,
  eval_DEGs = FALSE,
  eval_diff_abundance = FALSE,
  eval_batch_removal = FALSE,
  eval_trajectory = FALSE,
  eval_spatial_domain = FALSE
)
}
\arguments{
\item{SimulationResult}{A SCST object containing the simulated data.}

\item{eval_clustering}{Logical, whether to evaluate the clustering methods on the simulated data.}

\item{eval_cell_association}{Logical, whether to evaluate the methods for measuring the cell-cell association on the simulated data.}

\item{eval_annotation}{Logical, whether to evaluate the cell type annotation methods on the simulated data.}

\item{eval_DEGs}{Logical, whether to evaluate the methods for identifying DEGs on the simulated data.}

\item{eval_diff_abundance}{Logical, whether to evaluate the methods for testing differential abundance on the simulated data.}

\item{eval_batch_removal}{Logical, whether to evaluate the batch removal methods on the simulated data.}

\item{eval_trajectory}{Logical, whether to evaluate the methods of trajectory inference on the simulated data.}

\item{eval_spatial_domain}{Logical, whether to evaluate the methods for identifying spatial domains on the simulated spatial data.}
}
\description{
Evaluate Clustering Methods Using Simulated Data
}
