% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kersplat.R
\name{Kersplat_simulation}
\alias{Kersplat_simulation}
\title{Simulate ScRNA-seq Data Using Kersplat}
\usage{
Kersplat_simulation(
  estimated_result,
  cell_num = NULL,
  gene_num = NULL,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [SCST::Kersplat_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in Kersplat, see [splatter::KersplatParams()]}
}
\description{
Simulate ScRNA-seq Data Using Kersplat
}
