% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lun2.R
\name{Lun2_estimation}
\alias{Lun2_estimation}
\title{Estimation Parameters Using Lun2}
\usage{
Lun2_estimation(SCST_Object, min.size = 50, verbose = FALSE, ...)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{min.size}{Minimum size of clusters when identifying group of cells in the data.}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in Lun2, see [splatter::lun2Estimate()]}
}
\description{
Estimation Parameters Using Lun2
}
