% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lun2.R
\name{Lun2_simulation}
\alias{Lun2_simulation}
\title{Simulate ScRNA-seq Data Using Lun2}
\usage{
Lun2_simulation(
  estimated_result,
  cell_num = NULL,
  gene_num = NULL,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [splatter::Lun2_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in Lun2, see [splatter::Lun2Params()]}
}
\description{
Simulate ScRNA-seq Data Using Lun2
}
