% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lun.R
\name{Lun_estimation}
\alias{Lun_estimation}
\title{Estimation Parameters Using Lun}
\usage{
Lun_estimation(SCST_Object, verbose = FALSE)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}
}
\description{
Estimation Parameters Using Lun
}
