% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lun.R
\name{Lun_simulation}
\alias{Lun_simulation}
\title{Simulate ScRNA-seq Data Using Lun}
\usage{
Lun_simulation(
  estimated_result,
  cell_num = NULL,
  nGroups = 1,
  prop_group = 1,
  de_prop = 0.2,
  fc_group = 2,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [SCST::Lun_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{nGroups}{The expected number of groups in the simulated data. Default is 1.}

\item{prop_group}{The proportion of cells in each group. Default is 1.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{fc_group}{The fold change of DEGs between cell groups. Default is 2.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in Lun, see [splatter::LunParams()]}
}
\description{
Simulate ScRNA-seq Data Using Lun
}
