% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscat.R
\name{muscat_estimation}
\alias{muscat_estimation}
\title{Estimation Parameters Using muscat}
\usage{
muscat_estimation(
  SCST_Object,
  min_count = 0,
  min_cells = 0,
  min_genes = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{min_count, min_cells}{Only genes with a count > min_count in >= min_cells will be retained.}

\item{min_genes}{Only cells with a count > 0 in >= min_genes will be retained.}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in muscat, see [muscat::prepSim()]}
}
\description{
Estimation Parameters Using muscat
}
\references{
Crowell H L, Soneson C, Germain P L, et al. Muscat detects subpopulation-specific state transitions from multi-sample multi-condition single-cell transcriptomics data. Nature communications, 2020, 11(1): 1-12. <https://doi.org/10.1038/s41467-020-19894-4>
}
