% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscat.R
\name{muscat_simulation}
\alias{muscat_simulation}
\title{Simulate ScRNA-seq Data Using muscat}
\usage{
muscat_simulation(
  estimated_result,
  cell_num = NULL,
  gene_num = NULL,
  nGroups = 1,
  fc_group = 2,
  de_prop = 0.2,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [splatter::muscat_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{nGroups}{The expected number of groups in the simulated data. Default is 1.}

\item{fc_group}{The fold change of DEGs between cell groups. Default is 2.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in muscat, see [muscat::simData()]}
}
\description{
Simulate ScRNA-seq Data Using muscat
}
\references{
Crowell H L, Soneson C, Germain P L, et al. Muscat detects subpopulation-specific state transitions from multi-sample multi-condition single-cell transcriptomics data. Nature communications, 2020, 11(1): 1-12. <https://doi.org/10.1038/s41467-020-19894-4>
}
