% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign2.R
\name{scDesign2_estimation}
\alias{scDesign2_estimation}
\title{Estimation Parameters Using scDesign2}
\usage{
scDesign2_estimation(
  SCST_Object,
  cell_type = NULL,
  sim_method = "copula",
  marginal = "auto_choose",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{cell_type}{A character string specifying the cluster labels of cells.}

\item{sim_method}{Specification of the type of model. Default value is "copula", which selects the copula model. "ind" will select the (w/o copula) model.}

\item{marginal}{A string or a vector of strings of the marginal distribution. Must be one of "poisson", "nb", "zip", "zinb" or "gaussian". Default is "zinb".}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in scDesign2, see [scDesign2::fit_model_scDesign2()]}
}
\description{
Estimation Parameters Using scDesign2
}
\references{
Sun T, Song D, Li W V, et al. scDesign2: a transparent simulator that generates high-fidelity single-cell gene expression count data with gene correlations captured. Genome biology, 2021, 22(1): 1-37. <https://doi.org/10.1186/s13059-021-02367-2>
}
