% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign2.R
\name{scDesign2_simulation}
\alias{scDesign2_simulation}
\title{Simulate ScRNA-seq Data Using scDesign2}
\usage{
scDesign2_simulation(
  estimated_result,
  cell_num = NULL,
  prop_group = 1,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [SCST::scDesign2_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{prop_group}{The proportion of cells in each group. Default is 1.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in scDesign2, see [scDesign2::simulate_count_scDesign2()]}
}
\description{
Simulate ScRNA-seq Data Using scDesign2
}
\references{
Sun T, Song D, Li W V, et al. scDesign2: a transparent simulator that generates high-fidelity single-cell gene expression count data with gene correlations captured. Genome biology, 2021, 22(1): 1-37. <https://doi.org/10.1186/s13059-021-02367-2>
}
