% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign3.R
\name{scDesign3_estimation}
\alias{scDesign3_estimation}
\title{Estimation Parameters Using scDesign3}
\usage{
scDesign3_estimation(
  SCST_Object,
  family_use = "zinb",
  n_cores = 1,
  copula = "gaussian",
  path = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{family_use}{A string or a vector of strings of the marginal distribution. Must be one of "poisson", "nb", "zip", "zinb" or "gaussian". Default is "zinb".}

\item{n_cores}{An integer. The number of cores to use.}

\item{copula}{A string of the copula choice. Must be one of "gaussian" or "vine". Default is "gaussian".}

\item{path}{A Boolean symbol (TRUE or FLASE) which determines whether to simulate datasets with trajectory along the path. Default is false.}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in scDesign3, see [scDesign3::fit_marginal()], [scDesign3::fit_copula()], [scDesign3::extract_para()]}
}
\description{
Estimation Parameters Using scDesign3
}
\references{
Song, D., Wang, Q., Yan, G. et al. scDesign3 generates realistic in silico data for multimodal single-cell and spatial omics. Nat Biotechnol (2023). https://doi.org/10.1038/s41587-023-01772-1
}
