% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign3.R
\name{scDesign3_simulation}
\alias{scDesign3_simulation}
\title{Simulate ScRNA-seq Data Using scDesign3}
\usage{
scDesign3_simulation(
  estimated_result,
  n_cores = 1,
  family_use = "zinb",
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [scDesign3::fit_marginal()], [scDesign3::fit_copula()], [scDesign3::extract_para()]}

\item{n_cores}{An integer. The number of cores to use.}

\item{family_use}{A string or a vector of strings of the marginal distribution. Must be one of "poisson", "nb", "zip", "zinb" or "gaussian". Default is "zinb".}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in scDesign3, see [scDesign3::simu_new()]}
}
\description{
Simulate ScRNA-seq Data Using scDesign3
}
\references{
Song, D., Wang, Q., Yan, G. et al. scDesign3 generates realistic in silico data for multimodal single-cell and spatial omics. Nat Biotechnol (2023). https://doi.org/10.1038/s41587-023-01772-1
}
