% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDesign.R
\name{scDesign_simulation}
\alias{scDesign_simulation}
\title{Simulate ScRNA-seq Data Using scDesign}
\usage{
scDesign_simulation(
  SCST_Object,
  cell_num = NULL,
  nGroups = 1,
  prop_group = 1,
  de_prop = 0.2,
  fc_group = 2,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{SCST_Object}{The SCST object generated by [SCST::Create_SCST_Object()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{nGroups}{The expected number of groups in the simulated data. Default is 1.}

\item{prop_group}{The proportion of cells in each group. Default is 1.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{fc_group}{The fold change of DEGs between cell groups. Default is 2.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in scDesign, see [scDesign::design_data()]}
}
\description{
Simulate ScRNA-seq Data Using scDesign
}
\references{
Li W V, Li J J. A statistical simulator scDesign for rational scRNA-seq experimental design. Bioinformatics, 2019, 35(14): i41-i50. <https://doi.org/10.1093/bioinformatics/btz321>
}
