% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIP.R
\name{SCRIP_estimation}
\alias{SCRIP_estimation}
\title{Estimation Parameters Using SCRIP}
\usage{
SCRIP_estimation(SCST_Object, verbose = FALSE)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}
}
\description{
Estimation Parameters Using SCRIP
}
\references{
Qin F, Luo X, Xiao F, et al. SCRIP: an accurate simulator for single-cell RNA sequencing data. Bioinformatics, 2022, 38(5): 1304-1311. <https://doi.org/10.1093/bioinformatics/btab824>
}
