% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIP.R
\name{SCRIP_simulation}
\alias{SCRIP_simulation}
\title{Simulate ScRNA-seq Data Using SCRIP}
\usage{
SCRIP_simulation(
  estimated_result,
  mode = "GP-trendedBCV",
  cell_num = NULL,
  gene_num = NULL,
  nGroups = 1,
  prop_group = 1,
  de_prop = 0.2,
  de_facLoc = 0.1,
  de_facScale = 0.4,
  nBatches = 1,
  prop_batch = 1,
  batch_facLoc = 0.1,
  batch_facScale = 0.1,
  path = FALSE,
  verbose = FALSE,
  return_format,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [splatter::splatEstimate()]}

\item{mode}{SCRIP contains five different simulation modes, and you can specify which mode do you use (default is GP-trendedBCV): 1. GP-trendedBCV; 2. GP-commonBCV; 3. BGP-commonBCV 4. BP 5. BGP-trendedBCV}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{nGroups}{The expected number of groups in the simulated data. Default is 1.}

\item{prop_group}{The proportion of cells in each group. Default is 1.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{de_facLoc}{Location (meanlog) parameter for the differential expression factor log-normal distribution. Can be a vector.}

\item{de_facScale}{Scale (sdlog) parameter for the differential expression factor log-normal distribution. Can be a vector.}

\item{nBatches}{The number of cell batches to be simulated. Default is 1.}

\item{prop_batch}{The proportion of cells in each batch. Default is 1.}

\item{batch_facLoc}{Location (meanlog) parameter for the batch effect factor log-normal distribution. Can be a vector.}

\item{batch_facScale}{Scale (sdlog) parameter for the batch effect factor log-normal distribution. Can be a vector.}

\item{path}{A Boolean symbol (TRUE or FLASE) which determines whether to simulate datasets with trajectory along the path. Default is false.}

\item{verbose}{Whether to return messages or not}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{seed}{Random seed}

\item{...}{Other parameters represented in SCRIP, see [SCRIP::SCRIPsimu()]}
}
\description{
Simulate ScRNA-seq Data Using SCRIP
}
\references{
Qin F, Luo X, Xiao F, et al. SCRIP: an accurate simulator for single-cell RNA sequencing data. Bioinformatics, 2022, 38(5): 1304-1311. <https://doi.org/10.1093/bioinformatics/btab824>
}
