% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{SCST}
\alias{SCST}
\alias{SCST-class}
\title{The SCST Class}
\description{
S4 class that contains the settings and information used for estimating important parameters from real reference data
}
\section{Slots}{

\describe{
\item{\code{reference}}{A reference gene expression matrix}

\item{\code{methods}}{The method(s) used for simulation}

\item{\code{cell_num}}{The cell/spots/bin number of the reference matrix}

\item{\code{gene_num}}{The gene number of the reference matrix}

\item{\code{group_label}}{The group labels of cells/spots/bins}

\item{\code{batch_label}}{The batch labels of cells/spots/bins}

\item{\code{dilution_factor}}{If ERCC spike-in RNAs were added into the single-cell solution, the dilution factor can be used for estimating parameters from real using certain methods (such as SPARSim)}

\item{\code{volume}}{The volume of ERCC spike-in mix added into the single-cell solution}

\item{\code{spatial_x}}{The coordinates of the spots/bins in the spatial transcriptome data on the x-axis}

\item{\code{spatial_y}}{The coordinates of the spots/bins in the spatial transcriptome data on the y-axis}

\item{\code{seed}}{The random seed used for reproducibility}

\item{\code{estimation_result}}{The outcome of parameter estimation step for simulating new datasets}

\item{\code{estimation_time_memory}}{Time consumption and memory usage in the parameter estimation step}

\item{\code{estimation_parameters}}{The input parameters used in the estimation step}

\item{\code{simulation_result}}{The outcome of simulated data}

\item{\code{simulation_time_memory}}{Time consumption and memory usage in the data simulation step}

\item{\code{simulation_parameters}}{The input parameters used in the simulation step}

\item{\code{customed_setting}}{The prior settings of different simulation scenarios are specified for generating single-cell or ST datasets with different sizes, groups, DEGs, batches and trajectories.}
}}

