% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_datasets.R
\name{Simulate_datasets}
\alias{Simulate_datasets}
\title{Simulate Datasets}
\usage{
Simulate_datasets(SCST_Object, verbose = TRUE)
}
\arguments{
\item{SCST_Object}{A SCST object genaated by [SCST::Estimate_parameters] and [SCST::Set_customed_parameters] functions}

\item{verbose}{Whether to return messages or not}
}
\value{
A SCST object with all results of parameter estimation, simulation and used parameters.
}
\description{
Simulate Datasets
}
