% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim.R
\name{SPARSim_estimation}
\alias{SPARSim_estimation}
\title{Estimation Parameters Using SPARSim}
\usage{
SPARSim_estimation(SCST_Object, verbose = FALSE, ...)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in SRTsim, see [SPARSim::SPARSim_estimate_parameter_from_data()]}
}
\description{
Estimation Parameters Using SPARSim
}
\references{
Baruzzo G, Patuzzi I, Di Camillo B. SPARSim single cell: a count data simulator for scRNA-seq data[J]. Bioinformatics, 2020, 36(5): 1468-1475.
}
