% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splat.R
\name{Splat_estimation}
\alias{Splat_estimation}
\title{Estimation Parameters Using Splat}
\usage{
Splat_estimation(SCST_Object, verbose = FALSE, ...)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in Splat, see [splatter::splatEstimate()]}
}
\description{
Estimation Parameters Using Splat
}
\references{
Zappia L, Phipson B, Oshlack A. Splatter: simulation of single-cell RNA sequencing data[J]. Genome biology, 2017, 18(1): 174.
}
