% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTsim.R
\name{SRTsim_estimation}
\alias{SRTsim_estimation}
\title{Estimation Parameters Using SRTsim}
\usage{
SRTsim_estimation(
  SCST_Object,
  sim_schem = "domain",
  marginal = "auto_choose",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{sim_schem}{A character string specifying simulation scheme. "tissue" stands for tissue-based simulation; "domain" stands for domain-specific simulation. Default is "domain".}

\item{marginal}{Specification of the types of marginal distribution ("auto_choose", "zinb", "nb", "poisson", "zip")}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in SRTsim, see [SRTsim::srtsim_fit()]}
}
\description{
Estimation Parameters Using SRTsim
}
\references{
Zhu J, Shang L, Zhou X. SRTsim: spatial pattern preserving simulations for spatially resolved transcriptomics[J]. Genome Biology, 2023, 24(1): 39.
}
