% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTsim.R
\name{SRTsim_simulation}
\alias{SRTsim_simulation}
\title{Simulate ST Data Using SRTsim}
\usage{
SRTsim_simulation(estimated_result, verbose = FALSE, seed, return_format, ...)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [SCST::SRTsim_estimation()]}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{...}{Other parameters represented in SRTsim, see [SRTsim::srtsim_count()]}
}
\description{
Simulate ST Data Using SRTsim
}
\references{
Zhu J, Shang L, Zhou X. SRTsim: spatial pattern preserving simulations for spatially resolved transcriptomics[J]. Genome Biology, 2023, 24(1): 39.
}
