% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VeloSim.R
\name{VeloSim_estimation}
\alias{VeloSim_estimation}
\title{Estimation Parameters Using VeloSim}
\usage{
VeloSim_estimation(SCST_Object, verbose = FALSE)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}
}
\description{
Estimation Parameters Using VeloSim
}
\references{
Zhang Z, Zhang X. VeloSim: Simulating single cell gene-expression and RNA velocity. BioRxiv, 2021. <https://doi.org/10.1101/2021.01.11.426277>
}
