% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VeloSim.R
\name{VeloSim_simulation}
\alias{VeloSim_simulation}
\title{Simulate ScRNA-seq Data Using VeloSim}
\usage{
VeloSim_simulation(
  estimated_result,
  cell_num = NULL,
  gene_num = NULL,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters by [SCST::VeloSim_estimation()].}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{gene_num}{The expected number of simulated genes. If NULL, the original gene number in the reference data will be adopted.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in VeloSim, see [VeloSim::SimulateVeloTree()]}
}
\description{
Simulate ScRNA-seq Data Using VeloSim
}
\references{
Zhang Z, Zhang X. VeloSim: Simulating single cell gene-expression and RNA velocity. BioRxiv, 2021. <https://doi.org/10.1101/2021.01.11.426277>
}
