% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB-WaVE.R
\name{zinbwave_estimation}
\alias{zinbwave_estimation}
\title{Estimation Parameters Using ZINB-WaVE}
\usage{
zinbwave_estimation(SCST_Object, verbose = FALSE, ...)
}
\arguments{
\item{SCST_Object}{A SCST object generated by [SCST::Create_SCST_Object()]}

\item{verbose}{Whether to return messages or not}

\item{...}{Other parameters represented in ZINB-WaVE, see [splatter::ZINBParams()]}
}
\description{
Estimation Parameters Using ZINB-WaVE
}
