% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB-WaVE.R
\name{zinbwave_simulation}
\alias{zinbwave_simulation}
\title{Simulate ScRNA-seq Data Using ZINB-WaVE}
\usage{
zinbwave_simulation(
  estimated_result,
  cell_num = NULL,
  return_format,
  verbose = FALSE,
  seed,
  ...
)
}
\arguments{
\item{estimated_result}{The SCST object after estimating parameters using [splatter::zinbwave_estimation()]}

\item{cell_num}{The expected number of simulated cells. If NULL, the original cell number in the reference data will be adopted.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat.}

\item{verbose}{Whether to return messages or not}

\item{seed}{Random seed}

\item{...}{Other parameters represented in ZINB-WaVE, see [splatter::zinbSimulate()]}
}
\description{
Simulate ScRNA-seq Data Using ZINB-WaVE
}
