% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Set_customed_parameters}
\alias{Set_customed_parameters}
\title{Set Optional Parameters For Simulation}
\usage{
Set_customed_parameters(
  SCST_Object,
  return_format = "list",
  cell_num = NULL,
  gene_num = NULL,
  mode = NULL,
  nGroups = NULL,
  prop_group = NULL,
  de_prop = NULL,
  fc_group = NULL,
  de_facLoc = NULL,
  de_facScale = NULL,
  nBatches = NULL,
  prop_batch = NULL,
  batch_facLoc = NULL,
  batch_facScale = NULL,
  path = NULL
)
}
\arguments{
\item{SCST_Object}{The estimation result generated by [SCST::Estimate_parameters] function or the estimation function for an individual method.}

\item{return_format}{The format of returned simulation data. Choices: list, SingleCellExperiment and Seurat. Default is list.}

\item{cell_num}{The expected number of cells to be simulated.}

\item{gene_num}{The expected number of genes to be simulated.}

\item{mode}{SCRIP contains five different simulation modes, and you can specify which mode do you use (default is GP-trendedBCV): 1. GP-trendedBCV; 2. GP-commonBCV; 3. BGP-commonBCV 4. BP 5. BGP-trendedBCV}

\item{nGroups}{The expected number of cell groups to be simulated. Default is 1.}

\item{prop_group}{The proportion of cells in each group. Default is 1.}

\item{de_prop}{The proportion of DEGs over all genes between different simulated cell groups. Default is 0.2.}

\item{fc_group}{The fold change of the generated DEGs. Default is 2.}

\item{de_facLoc}{Location (meanlog) parameter for the differential expression factor log-normal distribution. Can be a vector.}

\item{de_facScale}{Scale (sdlog) parameter for the differential expression factor log-normal distribution. Can be a vector.}

\item{nBatches}{The expected number of cell batches to be simulated. Default is 1.}

\item{prop_batch}{The proportion of cells in each batch. Default is 1.}

\item{batch_facLoc}{Location (meanlog) parameter for the batch effect factor log-normal distribution. Can be a vector.}

\item{batch_facScale}{Scale (sdlog) parameter for the batch effect factor log-normal distribution. Can be a vector.}

\item{path}{A Boolean (TRUE or FLASE) which determines whether to simulate datasets with trajectory along the path. Default is false.}
}
\value{
A SCST object
}
\description{
Set Optional Parameters For Simulation
}
